requirements.txt
Flask
firebase-admin
openai
SpeechRecognition
pyttsx3
pydub
python-dotenv
requests